Separating the environment dynamics from the state of an environment instance (like current_state).
Having these two ideas in the same class was a bit messy before. This change was motivated by wanting to be able to take in a const Environment and be able to run trials. This wasn't possible when the Environment itself held the current state variables. Now, any number of Trial objects can be created from an Environment object. The latter simply representing the dynamics of an env.